Add changelog and requirements.txt
