+++ requirements.txt
@@ -1,0 +1,15 @@
+pyjwt
+cryptography
+requests
+coinbase
+attrs
+idna_ssl
+aiohttp
+async_timeout
+charset_normalizer
+async_timeout
+yarl
+psycopg2
+crypto_data
+dash
+plotly.express
